requirements.txt
absl-py==2.0.0
anyio @ file:///tmp/build/80754af9/anyio_1644481695334/work/dist
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///tmp/build/80754af9/argon2-cffi-bindings_1644553347904/work
astropy==6.0.0
astropy-iers-data==0.2024.2.26.0.28.55
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
attrs @ file:///croot/attrs_1695717823297/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///croot/beautifulsoup4-split_1681493039619/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
cached-property @ file:///tmp/build/80754af9/cached-property_1600785575025/work
cachetools==5.3.1
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
colorama @ file:///croot/colorama_1672386526460/work
comm @ file:///croot/comm_1671231121260/work
contourpy @ file:///opt/conda/conda-bld/contourpy_1663827406301/work
cupy-cuda11x==12.1.0
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython @ file:///home/conda/feedstock_root/build_artifacts/cython_1685025039295/work
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649908313000/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema @ file:///opt/conda/conda-bld/python-fastjsonschema_1661371079312/work
fastrlock==0.8.1
few==1.4.11
flatbuffers==23.5.26
fonttools==4.25.0
gast==0.5.4
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py @ file:///home/conda/feedstock_root/build_artifacts/h5py_1687263074382/work
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==6.8.0
ipykernel @ file:///croot/ipykernel_1691121631942/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1685727741709/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///croot/ipywidgets_1679394798311/work
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
Jinja2 @ file:///croot/jinja2_1666908132255/work
jsonschema @ file:///croot/jsonschema_1676558650973/work
jupyter @ file:///home/conda/feedstock_root/build_artifacts/jupyter_1670249595582/work
jupyter-console @ file:///croot/jupyter_console_1679999630278/work
jupyter-server @ file:///croot/jupyter_server_1671707632269/work
jupyter_client @ file:///croot/jupyter_client_1676329080601/work
jupyter_core @ file:///croot/jupyter_core_1679906564508/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///croot/jupyterlab_widgets_1679055282532/work
keras==2.14.0
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
libclang==16.0.6
LISAanalysistools==0.1
lisaconstants @ git+https://gitlab.in2p3.fr/lisa-simulation/constants.git@8245914b0791a79f4bb7606fadc2e0858e2b3f4b
lisainstrument @ git+https://gitlab.in2p3.fr/lisa-simulation/instrument.git@71f428e91c87afac781ff14e28253c223a1dbe7b
Markdown==3.4.4
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib @ file:///croot/matplotlib-suite_1679593461707/work
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mistune @ file:///home/builder/ci_310/mistune_1640791641754/work
ml-dtypes==0.2.0
munkres==1.1.4
nbclassic @ file:///croot/nbclassic_1681756170010/work
nbclient @ file:///tmp/build/80754af9/nbclient_1650290238894/work
nbconvert @ file:///opt/conda/conda-bld/nbconvert_1649751911790/work
nbformat @ file:///croot/nbformat_1694616755618/work
nest-asyncio @ file:///croot/nest-asyncio_1672387112409/work
notebook @ file:///croot/notebook_1690984815942/work
notebook_shim @ file:///croot/notebook-shim_1668160579331/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1687056313536/work
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.16.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging @ file:///croot/packaging_1693575174725/work
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.4.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
ply==3.11
pooch @ file:///croot/pooch_1695850093751/work
prometheus-client @ file:///tmp/abs_d3zeliano1/croots/recipe/prometheus_client_1659455100375/work
prompt-toolkit @ file:///croot/prompt-toolkit_1672387306916/work
protobuf==4.24.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyerfa==2.0.1.1
Pygments @ file:///croot/pygments_1684279966437/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5-sip==12.11.0
pyrsistent @ file:///home/builder/ci_310/pyrsistent_1640807196327/work
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
PyYAML==6.0.1
pyzmq @ file:///opt/conda/conda-bld/pyzmq_1657724186960/work
qtconsole @ file:///croot/qtconsole_1681394213385/work
QtPy @ file:///opt/conda/conda-bld/qtpy_1662014892439/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
requests-oauthlib==1.3.1
response==1.0.2
rsa==4.9
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy-split_1683900262125/work/base/dist/scipy-1.10.1-cp310-cp310-linux_x86_64.whl#sha256=b888affdeaa173fb719df0d12fc9d171eff704eaa2c14a5889e0f4bc5054d43e
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio @ file:///home/builder/ci_310/sniffio_1640794799774/work
soupsieve @ file:///croot/soupsieve_1696347547217/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorrt==8.5.3.1
termcolor==2.3.0
terminado @ file:///croot/terminado_1671751832461/work
testpath @ file:///opt/conda/conda-bld/testpath_1655908557405/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tornado @ file:///croot/tornado_1690848263220/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1677948868469/work
traitlets @ file:///croot/traitlets_1671143879854/work
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
urllib3==1.26.17
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client @ file:///home/builder/ci_310/websocket-client_1640795866898/work
Werkzeug==3.0.0
widgetsnbextension @ file:///croot/widgetsnbextension_1679313860248/work
wrapt==1.14.1
zipp==3.16.2
